Change type of total cost to In64 to handle high costs

+++ CompetitionCreator/Forms/Form1.cs
@@ -98,7 +98,7 @@ namespace CompetitionCreator
         }
         private void UpdateConflictCount()
         {
-            int conflicts = 0;
+            Int64 conflicts = 0;
             foreach (Constraint constraint in model.constraints)
             {
                 conflicts += constraint.conflict_cost;
